Allow selecting phases to run
There aren't any phases yet, but once we divide the mytster's repository
into phases, we'd want to run only select ones during develepment.

We differentiate "phases," which are the distinct building block of
generating a new mystery repository, and "steps," which are the actions
the player will have to follow on their way to the solution.