Step #1: Generate the main police report
Police reports are stored as commit messages on the `gtpd-archive`
branch. Their inconspicuous titles are just the police report ID, which
is randomly generated.

The main report is the one leading the player to the second phase. Right
now it's the only commit on the `gtpd-archive` branch, but we'll soon
add more commits before and after it to "hide" it in some "noise."

Right now the report's body is just some placeholder text. When we
implement the second phase we'll fill that in.