Add random paragraphs generator
To fill "fluff" commit messages we copy parts of Arthur Conan Doyle's
"The Adventues of Sherlock Holmes." According to Wikisource this work is
public domain because it was published prior to January 1st, 1925. So
using it should be fine.

The actual work of chaining paragraphs required some text manipulation.
First, we needed to dedent the entire thing. Then we dropped paragraphs
that started with a whitespace because those were story titles. Then we
drop chapter titles. Last we dropped the first 50 paragraphs to get rid
of the table of contents and parts of the exposition. This should
provide us with commit messages that seem more "natural."