Step #2: Fill access log with fake data
To cover for the suspects in the access log we generate a bunch of
random entries for other factory workers. This makes the second step
non-trivial, requiring the player to use either `git blame`, or better
yet `git log -S` to find who used the backdoor to the factory.

We store the names of the suspects with ROT13 to obfuscate the solution
in the source code.