Step #2: Mark suspects in factory access log
The main report points to the "cherry-picked" access log from the
factory. Right now only the suspects show up on the access log, but
we'll add more entries next to make the task non-trivial.

The suspects are entered to the log in reverse order to make sure that
the first suspect's interview on the 3rd step will (likely) be the first
one that the player will search to interview. That is because the normal
call to `git log -S <access point>` will output accesses from latest
(first entry) to earliest (last entry).