Step #1: Generate police reports on main's week
We previously generated police reports only before or after the week of
the main report, now we add reports on the same week to make things
a little bit harder. This will hopefully push the player to search for
commits only by the main detective and not any of the others.