Initialize the gitstery repository
The first thing to do is create a new repository to hold the mystery
data. We need to make sure all the git environment variables are cleared
so they won't mess with the script's operation (see comment in `cli` for
more information).

Generated a "mayor" user that creates the repository in a given date.