Step #4: Generate "investigation" texts
Two of the suspects aren't home, so the player will need to investigate
a bit. Investigation is done by reading the content of the address
commit. Right now it's pretty easy and would probably be done in the
same action as "interviewing" suspects, since `git show` will show the
diff as well. We'll complicate this a bit later.

The one suspect that does get interviewed provides a key clue to
catching the murderer. Investigating around the other suspects' houses
is the final step (for now).