Step #4: Add a layer of redirection to the investigation
We previously stored the "investigation" as a file in the house commit,
but that means that player will see it right away when they `git show`
the commit to read the commit message.

Instead, what we do is create an object that holds the investigation
text and the file points to that object. To actually display the
contents the player will have to `git show` the hash inside the file.

Because the blobs containing the text itself must be addressable or
otherwise `git clone` won't fetch them, we create a new orphan branch
that will hold all of them.