Step #1: Generate random reports around the main's week
Generate random reports with fluff information before and after the week
of the main report. This creates more commits on the `gtpd-branch` and
will push the player to use more flags in `git log` to find the correct
commit for the main report.

Right now we generate all the reports before or after the _week_ of the
main report, so a basic use of `git log --since` will lead the player to
the right commit. We'll fix that soon.